pyproject.toml
[tool.poetry]
name = "mybot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
python-dotenv = "^0.20.0"
pathlib = "^1.0.1"
Scrapy = "^2.6.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
ipykernel = "^6.13.0"
pytest = "^5.2"
pre-commit = "^2.17.0"
black = "21.12b0"        # code formatting
isort = "5.10.1"       # import sorting
flake8 = "4.0.1"       # code analysis

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
